Working Server scripts with Hananel model

+++ DGA/Pool.py
@@ -8,12 +8,20 @@ class Pool(dict):
     for key, value in self.items():
       subset_pool[key] = value    # Add all current items to subset pool (with condition)
 
-  def __setitem__(self, key, value):
+  def reinitialize_subsets(self):
     for sub_pool in self.subset_pools:
-      if key in sub_pool and not sub_pool.condition(key, value):
-        del sub_pool[key]         # If condition no longer met, delete from subset
-      elif sub_pool.condition(key, value):
-        sub_pool[key] = value     # Adjust sub pools to match change
+      for key, value in self.items():
+        sub_pool[key] = value     # Add all current items to subset pool (with condition)
+
+  def __setitem__(self, key, value):
+    # This is to avoid issues with loading Pool obj from file with pickle
+    # Constructor not called on load, so need to initialize subset_pools manually from Server.py
+    if 'subset_pools' in self.__dict__:
+      for sub_pool in self.subset_pools:
+        if key in sub_pool and not sub_pool.condition(key, value):
+          del sub_pool[key]         # If condition no longer met, delete from subset
+        elif sub_pool.condition(key, value):
+          sub_pool[key] = value     # Adjust sub pools to match change
     super().__setitem__(key, value)
 
   def __delitem__(self, key):
@@ -29,7 +37,7 @@ class Pool(dict):
   def update_subpools(self):
     for sub_pool in self.subset_pools:
       for key, value in self.items():
-        if not sub_pool.condition(key, value):
+        if not sub_pool.condition(key, value) and key in sub_pool:
           del sub_pool[key]         # If condition no longer met, delete from subset
         elif sub_pool.condition(key, value):
           sub_pool[key] = value     # Adjust sub pools to match change
@@ -40,5 +48,8 @@ class Subset_Pool(Pool):
     super().__init__()
 
   def __setitem__(self, key, value):
-    if self.condition(key, value):
-      super().__setitem__(key, value)   # Set item only if subset condition is met
+    # This is to avoid issues with loading Pool obj from file with pickle
+    # Constructor not called on load, so need to initialize subset_pools manually from Server.py
+    if 'condition' in self.__dict__:
+      if self.condition(key, value):
+        super().__setitem__(key, value)   # Set item only if subset condition is met
